+++ Source/WeenieObject.cpp
@@ -2985,7 +2985,7 @@ void CWeenieObject::CheckRegeneration(double rate, STypeAttribute2nd currentAttr
 			m_Qualities.InqAttribute(ENDURANCE_ATTRIBUTE, endurance, true);
 			float strAndEnd = (float)(strength + (endurance * 2));
 			float regenMod = 1.0 + (0.0494 * pow(strAndEnd, 1.179) / 100.0f); //formula deduced from values present in the client pdb.
-			regenMod = min(max(regenMod, 1.0), 2.1);
+			regenMod = min(max(regenMod, 1.0f), 2.1f);
 
 			rate *= regenMod;
 		}
@@ -4157,20 +4157,20 @@ float CWeenieObject::GetEffectiveArmorLevel(DamageEventData &damageData, bool bI
 		}
 		else
 		{
-			armorLevel = min(shieldSkill, armorLevel);
+			armorLevel = min((float)shieldSkill, armorLevel);
 		}
 	}
 
 	switch (damageData.damage_type)
 	{
-	case DAMAGE_TYPE::SLASH_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_SLASH_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0f)); break;
-	case DAMAGE_TYPE::PIERCE_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_PIERCE_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0f)); break;
-	case DAMAGE_TYPE::BLUDGEON_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_BLUDGEON_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0f)); break;
-	case DAMAGE_TYPE::FIRE_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_FIRE_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0f)); break;
-	case DAMAGE_TYPE::COLD_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_COLD_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0f)); break;
-	case DAMAGE_TYPE::ACID_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_ACID_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0f)); break;
-	case DAMAGE_TYPE::ELECTRIC_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_ELECTRIC_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0f)); break;
-	case DAMAGE_TYPE::NETHER_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_NETHER_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0f)); break;
+	case DAMAGE_TYPE::SLASH_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_SLASH_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0)); break;
+	case DAMAGE_TYPE::PIERCE_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_PIERCE_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0)); break;
+	case DAMAGE_TYPE::BLUDGEON_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_BLUDGEON_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0)); break;
+	case DAMAGE_TYPE::FIRE_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_FIRE_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0)); break;
+	case DAMAGE_TYPE::COLD_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_COLD_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0)); break;
+	case DAMAGE_TYPE::ACID_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_ACID_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0)); break;
+	case DAMAGE_TYPE::ELECTRIC_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_ELECTRIC_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0)); break;
+	case DAMAGE_TYPE::NETHER_DAMAGE_TYPE: armorLevel *= max(0.0, min(InqFloatQuality(ARMOR_MOD_VS_NETHER_FLOAT, 1.0f, bIgnoreMagicArmor), 2.0)); break;
 	}
 
 	if (isShield)
@@ -4271,7 +4271,7 @@ void CWeenieObject::TakeDamage(DamageEventData &damageData)
 			else
 				resistanceNatural = 1.0f - (((0.1666667 * strAndEnd) - 23.33333) / 100.f);
 
-			resistanceNatural = max(resistanceNatural, 0.5);
+			resistanceNatural = max(resistanceNatural, 0.5f);
 
 			//natural resistances only work if they beat the buffed value before any debuffs are applied.
 			if (resistanceNatural < buffDetails.enchantedValue_DecreasingOnly)
@@ -5992,7 +5992,7 @@ DWORD CWeenieObject::GetEffectiveManaConversionSkill()
 
 		if (manaMod > 0.0)
 		{
-			manaConvSkill = (DWORD)max(0, ((int)manaConvSkill * (1.0 + manaMod)));
+			manaConvSkill = (DWORD)max(0, (int)(manaConvSkill * (1.0 + manaMod)));
 		}
 	}
 
@@ -6140,7 +6140,7 @@ bool CWeenieObject::TryMeleeEvade(DWORD attackSkill)
 					DWORD endurance = 0;
 					m_Qualities.InqAttribute(ENDURANCE_ATTRIBUTE, endurance, true);
 					float noStaminaUseChance = ((float)endurance - 100.0) / 400.0; //made up formula: 75% reduction at 400 endurance.
-					noStaminaUseChance = min(max(noStaminaUseChance, 0.0), 0.75);
+					noStaminaUseChance = min(max(noStaminaUseChance, 0.0f), 0.75f);
 					if (Random::RollDice(0.0, 1.0) > noStaminaUseChance)
 						AdjustStamina(-1); // failed the roll, use stamina.
 				}
@@ -6226,7 +6226,7 @@ bool CWeenieObject::TryMissileEvade(DWORD attackSkill)
 				DWORD endurance = 0;
 				m_Qualities.InqAttribute(ENDURANCE_ATTRIBUTE, endurance, true);
 				float noStaminaUseChance = ((float)endurance - 100.0) / 400.0; //made up formula: 75% reduction at 400 endurance.
-				noStaminaUseChance = min(max(noStaminaUseChance, 0.0), 0.75);
+				noStaminaUseChance = min(max(noStaminaUseChance, 0.0f), 0.75f);
 				if (Random::RollDice(0.0, 1.0) > noStaminaUseChance)
 					AdjustStamina(-1); // failed the roll, use stamina.
 			}
